Options to allow the usage of commercial version of Observium
Added some extra variables to allow the usage of the paid for version of
Observium. It will require a username and password and the overriding of
observium_commercial variable to true. This will then use the commercial
version rather than the default community version.

No in place upgrades are performed.